Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 int main(){
-    float n;
+    double n;
     scanf("%f",&n);
     printf("You entered: %.4f\n",n);
     return 0;
